+++ CHANGELOG.md
@@ -59,6 +59,7 @@ end
 * `pay_customer` now supports a `default_payment_processor` option to automatically create a Pay::Customer record - @excid3
 * Added `enabled_processors` to Pay config. This lets you choose exactly which processors will be enabled. - @cjilbert504
 * Add `sync!` method to Pay::Subscription instances - @excid3
+* Ignore Stripe charges that don't have a customer ID - @excid3
 
 ### 3.0.24